pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "bloattables"
version = "0.1.0"
authors = [
    {name = "Elliot Simpson", email = "elliot@p-s.co.nz"},
]
description = "Simple library for creating and uploading test data to the cloud."
readme = "README.rst"
requires-python = ">=3.12"
keywords = ["cloud", "aws", "data"]
license = {text = "GNU AGPL"}
classifiers = [
      'Intended Audience :: Developers',
      'Programming Language :: Python :: 3',
]
dependencies = [
    "numpy==2.2.3",
    "pandas==2.2.3",
    "pandera==0.22.1",
    "boto3==1.36.21",
    "botocore==1.36.21",
    "python-dotenv>=1.0.1",
    "pyathena>=3.12.2",
    "pyarrow>=19.0.0",
    "ruff>=0.9.6",
    "google-cloud-storage>=3.0.0",
]

[project.optional-dependencies]
dev = ["build"]

[project.urls]
"Homepage" = "https://github.com/ByteCampNZ/bloattables"

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]

[dependency-groups]
dev = [
    "boto3-stubs>=1.36.21",
    "pandas-stubs>=2.2.3.241126",
    "ruff>=0.9.6",
]
